split client
