+++ Cargo.toml
@@ -6,5 +6,6 @@ members = [
     "futurecommander_representation",
     "futurecommander_daemon",
     "futurecommander_gui",
-    "futurecommander_protocol"
+    "futurecommander_protocol",
+    "futurecommander_client"
 ]
