+++ futurecommander_client/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "futurecommander_client"
+version = "0.1.0"
+authors = ["François Cadeillan <francois@azsystem.fr>"]
+edition = "2018"
+license = "GPL-3.0-or-later"
+
+[dependencies]
+futurecommander_protocol = { version = "0.1.0", path = "../futurecommander_protocol" }
+tokio = "0.1.21"
